pyproject.toml
[project]
name = "papi"
version = "0.0.1"
description = "pAPI – Python/Pluggable API Framework \n **pAPI** (short for **Python API** or **Pluggable API**) is a modular micro-framework built on top of [FastAPI](https://fastapi.tiangolo.com/), designed to simplify the creation of composable and interoperable web APIs. \n It extends FastAPI’s routing system to enable advanced features such as exposing HTTP endpoints as MCP tools, making it a powerful choice for agent-based applications, modular services, and programmable API ecosystems."

authors = [
    { name = "Eduardo Firvida", email = "efirvida@gmail.com" }
]

dependencies = [
    "aioredis>=2.0.1",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "beanie>=1.29.0",
    "click-default-group>=1.2.4",
    "fastapi>=0.115.12",
    "fastmcp>=2.5.1",
    "filetype>=1.2.0",
    "ipython>=9.2.0",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "python-arango-async>=0.0.3",
    "pyyaml>=6.0.2",
    "redis>=6.1.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.2",
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.11.6",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.29.1",
    "mkdocs-autorefs>=1.4.2",
    "mdx-include>=1.4.2",
    "markdown-include-variants>=0.0.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["papi"]
